printout access token
`spark config identify` will show the access token used as well for users wanting to quickly grab one for testing. 

Might be better to clean up this portion of code in future to list out all the parameters in the profile than to explicitly use `console.log` for each specific param!