refactor listAccessTokens to reduce promises
The documentation for when.js suggests that "when.defer()" is
discouraged and we should use when.{promise, lift, try} instead in most
cases.

I also realized that *most* of places that return promises are
unnecessary, since the things they're returning are *already* promises
up the chain, so I reduced the noise there.