tokens: implement new token command
Implement a command to create a new access token on your account.
Requires your username and password -- like all access token commands --
and optionally takes the "username" for the token (shown as the "client"
when listing tokens).

The "password" for access tokens appears to be unused currently, so the
new code handles this the same way as the login API call, using the
string "client_secret_here" as a placeholder.