Stop telling users to run `sudo brew`
It's not necessary and will ultimately cause users grief (see https://github.com/Homebrew/homebrew/issues/39701). We don't support it.

Also, libusb will be installed as a dependency of dfu-util, so users don't need to do so manually.

I'm not interested in signing a CLA, so please feel free to apply these changes manually.