Prompt to install UDEV rules for dfu
On Linux only root can access USB devices through dfu. Right now doing
`particle flash --usb` will say "no dfu device found" without suggesting
how to fix the issue.

It's possible to use sudo to work around this, but a better solution is
to install UDEV rules to allow any user to access a particular device.

This PR prompts the user to install UDEV rules when the dfu-util failure
indicates missing permissions.

The UDEV rules was checked for the Photon on Ubuntu 15.10.

I don't have a Core on hand so I can't check what is the USB product ID
for the Core in DFU mode. Can somebody find that out? `lsub | grep 1d50`
in Linux when the Core is in DFU mode should do it.

Somebody working on the CLI should review the code and check if I put
the code in the right place as I'm not familiar with the structure of
the application.