Use child_process.spawn to get verbose DFU output.
This fixes the issue of DFU output not showing for write operations.
The trick to making it work with `child_process.spawn` was to use
`sudo` as the command and `dfu-util` as the first arg when
`useSudoForDfu` is set to true. This is because `spawn` is expecting a
command and arguments, *not* a command line like `exec`.

Fixes: #80