Use serialNumber to identify serial devices on Linux
On Linux Ubuntu 15.10, the reported vendor ID and product ID are the bus IDs instead of the device IDs. As a backup to the VID and PID identification, do a string matching on the `serialNumber` reported by the `serialport` module.

The expected strings for the Core, Photon and Electron are:
```
serialNumber: 'Spark_Devices_Spark_Core_with_WiFi_6D7238764957',
serialNumber: 'Particle_Photon_with_WiFi_00000000050C',
serialNumber: 'Particle_Electron_00000000050C',
```

Fixes #190